Adjusted GPS configuration code.
